refactor traits out of distance module;
